split off swap methods
