+++ dango/dex/src/core.rs
@@ -1,3 +1,4 @@
+pub mod geometric;
 mod liquidity_pool;
 mod market_order;
 mod merged_orders;
@@ -5,6 +6,7 @@ mod order_filling;
 mod order_matching;
 mod router;
 mod types;
+pub mod xyk;
 
 pub use {
     liquidity_pool::*, market_order::*, merged_orders::*, order_filling::*, order_matching::*,
