+++ dango/dex/src/core/liquidity_pool.rs
@@ -1,13 +1,11 @@
 use {
+    crate::core::{geometric, xyk},
     anyhow::ensure,
     dango_oracle::OracleQuerier,
     dango_types::dex::{PairParams, PassiveLiquidity},
     grug::{
-        Bounded, Coin, CoinPair, Denom, Inner, IsZero, MathResult, MultiplyFraction, MultiplyRatio,
-        Number, NumberConst, StdResult, Udec128, Uint128, ZeroExclusiveOneExclusive,
-        ZeroExclusiveOneInclusive,
+        Coin, CoinPair, Denom, MultiplyFraction, Number, NumberConst, StdResult, Udec128, Uint128,
     },
-    std::{cmp, iter},
 };
 
 const HALF: Udec128 = Udec128::new_percent(50);
@@ -143,16 +141,16 @@ impl PassiveLiquidityPool for PairParams {
     ) -> anyhow::Result<(CoinPair, Uint128)> {
         let mint_ratio = match (&self.pool_type, lp_token_supply) {
             (PassiveLiquidity::Xyk { .. }, Uint128::ZERO) => {
-                return xyk_add_initial_liquidity(reserve, deposit.clone());
+                return xyk::add_initial_liquidity(reserve, deposit.clone());
             },
             (PassiveLiquidity::Geometric { .. }, Uint128::ZERO) => {
-                return geometric_add_initial_liquidity(reserve, deposit.clone());
+                return geometric::add_initial_liquidity(reserve, deposit.clone());
             },
             (PassiveLiquidity::Xyk { .. }, _) => {
-                xyk_add_subsequent_liquidity(&mut reserve, deposit.clone())?
+                xyk::add_subsequent_liquidity(&mut reserve, deposit.clone())?
             },
             (PassiveLiquidity::Geometric { .. }, _) => {
-                geometric_add_subsequent_liquidity(oracle_querier, &mut reserve, deposit.clone())?
+                geometric::add_subsequent_liquidity(oracle_querier, &mut reserve, deposit.clone())?
             },
         };
 
@@ -196,24 +194,13 @@ impl PassiveLiquidityPool for PairParams {
         let output_reserve = reserve.amount_of(&output_denom)?;
 
         let output_amount_after_fee = match self.pool_type {
-            PassiveLiquidity::Xyk { .. } => {
-                // Solve A * B = (A + input_amount) * (B - output_amount) for output_amount
-                // => output_amount = B - (A * B) / (A + input_amount)
-                // Round so that user takes the loss.
-                let output_amount =
-                    output_reserve.checked_sub(input_reserve.checked_multiply_ratio_ceil(
-                        output_reserve,
-                        input_reserve.checked_add(input.amount)?,
-                    )?)?;
-
-                // Apply swap fee. Round so that user takes the loss.
-                output_amount
-                    .checked_mul_dec_floor(Udec128::ONE - self.swap_fee_rate.into_inner())?
-            },
-            PassiveLiquidity::Geometric { .. } => {
-                // TODO: implement
-                todo!()
-            },
+            PassiveLiquidity::Xyk { .. } => xyk::swap_exact_amount_in(
+                input.amount,
+                input_reserve,
+                output_reserve,
+                self.swap_fee_rate,
+            )?,
+            PassiveLiquidity::Geometric { .. } => geometric::swap_exact_amount_in()?,
         };
 
         let output = Coin {
@@ -248,29 +235,13 @@ impl PassiveLiquidityPool for PairParams {
         );
 
         let input_amount = match self.pool_type {
-            PassiveLiquidity::Xyk { .. } => {
-                // Apply swap fee. In SwapExactIn we multiply ask by (1 - fee) to get the
-                // offer amount after fees. So in this case we need to divide ask by (1 - fee)
-                // to get the ask amount after fees.
-                // Round so that user takes the loss.
-                let output_amount_before_fee = output
-                    .amount
-                    .checked_div_dec_ceil(Udec128::ONE - self.swap_fee_rate.into_inner())?;
-
-                // Solve A * B = (A + input_amount) * (B - output_amount) for input_amount
-                // => input_amount = (A * B) / (B - output_amount) - A
-                // Round so that user takes the loss.
-                Uint128::ONE
-                    .checked_multiply_ratio_floor(
-                        input_reserve.checked_mul(output_reserve)?,
-                        output_reserve.checked_sub(output_amount_before_fee)?,
-                    )?
-                    .checked_sub(input_reserve)?
-            },
-            PassiveLiquidity::Geometric { .. } => {
-                // TODO: implement
-                todo!()
-            },
+            PassiveLiquidity::Xyk { .. } => xyk::swap_exact_amount_out(
+                output.amount,
+                input_reserve,
+                output_reserve,
+                self.swap_fee_rate,
+            )?,
+            PassiveLiquidity::Geometric { .. } => geometric::swap_exact_amount_out()?,
         };
 
         let input = Coin {
@@ -296,7 +267,7 @@ impl PassiveLiquidityPool for PairParams {
         let quote_reserve = reserve.amount_of(&quote_denom)?;
 
         match self.pool_type {
-            PassiveLiquidity::Xyk { order_spacing } => xyk_reflect_curve(
+            PassiveLiquidity::Xyk { order_spacing } => xyk::reflect_curve(
                 base_reserve,
                 quote_reserve,
                 order_spacing,
@@ -305,7 +276,7 @@ impl PassiveLiquidityPool for PairParams {
             PassiveLiquidity::Geometric {
                 ratio,
                 order_spacing,
-            } => geometric_reflect_curve(
+            } => geometric::reflect_curve(
                 base_reserve,
                 quote_reserve,
                 ratio,
@@ -314,263 +285,6 @@ impl PassiveLiquidityPool for PairParams {
             ),
         }
     }
-}
-
-fn xyk_add_initial_liquidity(
-    mut reserve: CoinPair,
-    deposit: CoinPair,
-) -> anyhow::Result<(CoinPair, Uint128)> {
-    reserve.merge(deposit)?;
-
-    // TODO: apply a scaling factor? e.g. 1,000,000 LP tokens per unit of invariant.
-    let mint_amount = xyk_normalized_invariant(&reserve)?;
-
-    Ok((reserve, mint_amount))
-}
-
-// TODO: use oracle price
-fn geometric_add_initial_liquidity(
-    mut reserve: CoinPair,
-    deposit: CoinPair,
-) -> anyhow::Result<(CoinPair, Uint128)> {
-    reserve.merge(deposit)?;
-
-    // TODO: apply a scaling factor? e.g. 1,000,000 LP tokens per unit of invariant.
-    let mint_amount = xyk_normalized_invariant(&reserve)?;
-
-    Ok((reserve, mint_amount))
-}
-
-fn xyk_add_subsequent_liquidity(
-    reserve: &mut CoinPair,
-    deposit: CoinPair,
-) -> anyhow::Result<Udec128> {
-    let invariant_before = xyk_normalized_invariant(&reserve)?;
-
-    // Add the used funds to the pool reserves.
-    reserve.merge(deposit)?;
-
-    // Compute the proportional increase in the invariant.
-    let invariant_after = xyk_normalized_invariant(&reserve)?;
-    let invariant_ratio = Udec128::checked_from_ratio(invariant_after, invariant_before)?;
-
-    // Compute the mint ratio from the invariant ratio based on the curve type.
-    // This ensures that an unbalances provision will be equivalent to a swap
-    // followed by a balancedliquidity provision.
-    Ok(invariant_ratio.checked_sub(Udec128::ONE)?)
-}
-
-fn geometric_add_subsequent_liquidity(
-    oracle_querier: &mut OracleQuerier,
-    reserve: &mut CoinPair,
-    deposit: CoinPair,
-) -> anyhow::Result<Udec128> {
-    fn oracle_value(
-        oracle_querier: &mut OracleQuerier,
-        coin_pair: &CoinPair,
-    ) -> anyhow::Result<Udec128> {
-        let first_value = oracle_querier
-            .query_price(coin_pair.first().denom, None)?
-            .value_of_unit_amount(*coin_pair.first().amount)?;
-        let second_value = oracle_querier
-            .query_price(coin_pair.second().denom, None)?
-            .value_of_unit_amount(*coin_pair.second().amount)?;
-        Ok(first_value.checked_add(second_value)?)
-    }
-
-    let deposit_value = oracle_value(oracle_querier, &deposit)?;
-    let reserve_value = oracle_value(oracle_querier, &reserve)?;
-
-    Ok(deposit_value.checked_div(reserve_value.checked_add(deposit_value)?)?)
-}
-
-fn xyk_reflect_curve(
-    base_reserve: Uint128,
-    quote_reserve: Uint128,
-    order_spacing: Udec128,
-    swap_fee_rate: Bounded<Udec128, ZeroExclusiveOneExclusive>,
-) -> StdResult<(
-    Box<dyn Iterator<Item = (Udec128, Uint128)>>,
-    Box<dyn Iterator<Item = (Udec128, Uint128)>>,
-)> {
-    // Compute the marginal price. We will place orders above and below this price.
-    let marginal_price = Udec128::checked_from_ratio(quote_reserve, base_reserve)?;
-
-    // Construct the bid order iterator.
-    // Start from the marginal price minus the swap fee rate.
-    let one_sub_fee_rate = Udec128::ONE.checked_sub(*swap_fee_rate)?;
-    let mut maybe_price = marginal_price.checked_mul(one_sub_fee_rate).ok();
-    let mut prev_size = Uint128::ZERO;
-    let mut prev_size_quote = Uint128::ZERO;
-    let bids = iter::from_fn(move || {
-        // Terminate if price is less or equal to zero.
-        let price = match maybe_price {
-            Some(price) if price.is_non_zero() => price,
-            _ => return None,
-        };
-
-        // Compute the total order size (in base asset) at this price.
-        let quote_reserve_div_price = quote_reserve.checked_div_dec(price).ok()?;
-        let mut size = quote_reserve_div_price.checked_sub(base_reserve).ok()?;
-
-        // Compute the order size (in base asset) at this price.
-        //
-        // This is the difference between the total order size at
-        // this price, and that at the previous price.
-        let mut amount = size.checked_sub(prev_size).ok()?;
-
-        // Compute the total order size (in quote asset) at this price.
-        let mut amount_quote = amount.checked_mul_dec_ceil(price).ok()?;
-        let mut size_quote = prev_size_quote.checked_add(amount_quote).ok()?;
-
-        // If total order size (in quote asset) is greater than the
-        // reserve, cap it to the reserve size.
-        if size_quote > quote_reserve {
-            size_quote = quote_reserve;
-            amount_quote = size_quote.checked_sub(prev_size_quote).ok()?;
-            amount = amount_quote.checked_div_dec_floor(price).ok()?;
-            size = prev_size.checked_add(amount).ok()?;
-        }
-
-        // If order size is zero, we have ran out of liquidity.
-        // Terminate the iterator.
-        if amount.is_zero() {
-            return None;
-        }
-
-        // Update the iterator state.
-        prev_size = size;
-        prev_size_quote = size_quote;
-        maybe_price = price.checked_sub(order_spacing).ok();
-
-        Some((price, amount))
-    });
-
-    // Construct the ask order iterator.
-    let one_plus_fee_rate = Udec128::ONE.checked_add(*swap_fee_rate)?;
-    let mut maybe_price = marginal_price.checked_mul(one_plus_fee_rate).ok();
-    let mut prev_size = Uint128::ZERO;
-    let asks = iter::from_fn(move || {
-        let price = maybe_price?;
-
-        // Compute the total order size (in base asset) at this price.
-        let quote_reserve_div_price = quote_reserve.checked_div_dec(price).ok()?;
-        let size = base_reserve.checked_sub(quote_reserve_div_price).ok()?;
-
-        // If total order size (in base asset) exceeds the base asset
-        // reserve, cap it to the reserve size.
-        let size = cmp::min(size, base_reserve);
-
-        // Compute the order size (in base asset) at this price.
-        //
-        // This is the difference between the total order size at
-        // this price, and that at the previous price.
-        let amount = size.checked_sub(prev_size).ok()?;
-
-        // If order size is zero, we have ran out of liquidity.
-        // Terminate the iterator.
-        if amount.is_zero() {
-            return None;
-        }
-
-        // Update the iterator state.
-        prev_size = size;
-        maybe_price = price.checked_add(order_spacing).ok();
-
-        Some((price, amount))
-    });
-
-    Ok((Box::new(bids), Box::new(asks)))
-}
-
-fn geometric_reflect_curve(
-    mut base_reserve: Uint128,
-    mut quote_reserve: Uint128,
-    ratio: Bounded<Udec128, ZeroExclusiveOneInclusive>,
-    order_spacing: Udec128,
-    swap_fee_rate: Bounded<Udec128, ZeroExclusiveOneExclusive>,
-) -> StdResult<(
-    Box<dyn Iterator<Item = (Udec128, Uint128)>>,
-    Box<dyn Iterator<Item = (Udec128, Uint128)>>,
-)> {
-    // FIXME: use oracle price instead
-    let marginal_price = Udec128::checked_from_ratio(quote_reserve, base_reserve)?;
-
-    // Construct bid price iterator with decreasing prices.
-    let bids = {
-        let one_sub_fee_rate = Udec128::ONE.checked_sub(*swap_fee_rate)?;
-        let bid_starting_price = marginal_price.checked_mul(one_sub_fee_rate)?;
-        let mut maybe_price = Some(bid_starting_price);
-
-        let bid_prices = iter::from_fn(move || {
-            let price = match maybe_price {
-                Some(price) if price.is_non_zero() => price,
-                _ => return None,
-            };
-
-            maybe_price = price.checked_sub(order_spacing).ok();
-
-            Some(price)
-        });
-
-        let bid_sizes_in_quote = iter::from_fn(move || {
-            let size = match quote_reserve.checked_mul_dec(ratio.into_inner()) {
-                Ok(size_in_quote) => size_in_quote,
-                Err(_) => return None,
-            };
-
-            quote_reserve.checked_sub_assign(size).ok()?;
-
-            Some(size)
-        });
-
-        bid_prices
-            .zip(bid_sizes_in_quote)
-            .filter_map(|(price, size_in_quote)| {
-                let size = size_in_quote.checked_div_dec_floor(price).ok()?;
-                Some((price, size))
-            })
-    };
-
-    // Construct ask price iterator with increasing prices.
-    let asks = {
-        let one_plus_fee_rate = Udec128::ONE.checked_add(*swap_fee_rate)?;
-        let ask_starting_price = marginal_price.checked_mul(one_plus_fee_rate)?;
-        let mut maybe_price = Some(ask_starting_price);
-
-        let ask_prices = iter::from_fn(move || {
-            let price = match maybe_price {
-                Some(price) if price.is_non_zero() => price,
-                _ => return None,
-            };
-            maybe_price = price.checked_add(order_spacing).ok();
-            Some(price)
-        });
-
-        // Construct ask size placing `ratio` of the remaining liquidity of
-        // base reserve into each order.
-        let ask_sizes = iter::from_fn(move || {
-            let size = match base_reserve.checked_mul_dec(ratio.into_inner()) {
-                Ok(size) => size,
-                Err(_) => return None,
-            };
-            base_reserve.checked_sub_assign(size).ok()?;
-            Some(size)
-        });
-
-        ask_prices.zip(ask_sizes)
-    };
-
-    Ok((Box::new(bids), Box::new(asks)))
-}
-
-/// Compute `sqrt(A * B)`, where `A` and `B` are the reserve amount of the two
-/// assets in an xyk pool.
-fn xyk_normalized_invariant(reserve: &CoinPair) -> MathResult<Uint128> {
-    let a = *reserve.first().amount;
-    let b = *reserve.second().amount;
-
-    a.checked_mul(b)?.checked_sqrt()
 }
 
 /// Compute `|a - b|`.
@@ -589,7 +303,7 @@ mod tests {
     use {
         super::*,
         dango_types::constants::{eth, usdc},
-        grug::{Bounded, Coins, coins},
+        grug::{Bounded, Coins, Inner, coins},
         test_case::test_case,
     };
